Edit folder structure
