+++ src/extension.ts
@@ -1,11 +1,11 @@
 import * as vscode from "vscode";
-import * as starter from "./util/starter";
+import { generateProject } from "./commands/generateProject";
 
 // this method is called when the extension is activated
 // the extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext): void {
   context.subscriptions.push(
-    vscode.commands.registerCommand("extension.microProfileStarter", starter.generateProject)
+    vscode.commands.registerCommand("extension.microProfileStarter", generateProject)
   );
 }
